requirements.txt
ajsonrpc==1.2.0
altair==5.5.0
anyio==4.6.2
attrs==24.2.0
blinker==1.9.0
bottle==0.13.2
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
fonttools==4.47.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.26.3
opencv-python==4.9.0.80
oracledb==3.1.1
packaging==23.2
paho-mqtt==2.1.0
pandas==2.1.4
Pillow==10.2.0
plotly==5.18.0
protobuf==4.25.2
pyarrow==14.0.2
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
scikit-learn==1.4.0
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
streamlit==1.31.1
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.2.0
torchvision==0.17.0
tornado==6.4
typing_extensions>=4.10.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.1.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
